+++ generator-core/src/test/java/org/apiaddicts/apitools/apigen/generatorcore/generator/implementations/java/apigen/web/controller/endpoints/PostParentChildEndpointBuilderTests.java
@@ -18,93 +18,84 @@ import org.mockito.Mockito;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
-class PutEndpointBuilderTests {
+class PostParentChildEndpointBuilderTests {
 
     static TypeSpec typeSpec;
 
     @BeforeAll
     static void init() {
-        Endpoint endPoint = EndpointObjectMother.standardPut("put", "EntityName");
+        Endpoint endPoint = EndpointObjectMother.standardParentChildPost("postParentChild", "Child");
         JavaEntitiesData entitiesData = Mockito.mock(JavaEntitiesData.class);
         ApigenContext ctx = ApigenContextObjectMother.create();
         ctx.setEntitiesData(entitiesData);
-        PutEndpointBuilder<ApigenContext>
-                builderEndpoint = new PutEndpointBuilder<>(new Mapping("/entities"), endPoint, ctx,
+        PostParentChildEndpointBuilder<ApigenContext>
+                builderEndpoint = new PostParentChildEndpointBuilder<>(new Mapping("/parents/{id}/children"), endPoint, ctx,
                 ConfigurationObjectMother.create());
-        TypeSpec.Builder builder = TypeSpec.classBuilder("PutEndpoint");
+        TypeSpec.Builder builder = TypeSpec.classBuilder("PostEndpointParentChild");
         builderEndpoint.apply(builder);
         typeSpec = builder.build();
     }
 
     @Test
-    void givenPutEndpointBuilder_whenBuild_thenNameCorrect() {
-        assertEquals("PutEndpoint", typeSpec.name);
+    void givenPostEndpointBuilder_whenBuild_thenNameCorrect() {
+        assertEquals("PostEndpointParentChild", typeSpec.name);
     }
 
     @Test
-    void givenPutEndpointBuilder_whenBuild_thenModifierIsPublic() {
+    void givenPostEndpointBuilder_whenBuild_thenNoHaveModifier() {
         assertEquals(0, typeSpec.modifiers.size());
     }
 
     @Test
-    void givenPutEndpointBuilder_whenBuild_thenKindIsClass() {
+    void givenPostEndpointBuilder_whenBuild_thenKindIsClass() {
         assertEquals("CLASS", typeSpec.kind.toString());
     }
 
     @Test
-    void givenPutEndpointBuilder_whenBuild_thenNoHaveAnnotation() {
+    void givenPostEndpointBuilder_whenBuild_thenNoHaveAnnotation() {
         assertEquals(0, typeSpec.annotations.size());
     }
 
     @Test
-    void givenPutEndpointBuilder_whenBuild_thenNoHaveModifier() {
-        assertEquals(0, typeSpec.modifiers.size());
-    }
-
-    @Test
-    void givenPutEndpointBuilder_whenBuild_thenNoHaveFieldsSpecs() {
+    void givenPostEndpointBuilder_whenBuild_thenNoHaveFieldsSpecs() {
         assertEquals(0, typeSpec.fieldSpecs.size());
     }
 
     @Test
-    void givenPutEndpointBuilder_whenBuild_thenHaveSuperClassAndIsCorrect() {
+    void givenPostEndpointBuilder_whenBuild_thenHaveSuperClassAndIsCorrect() {
         assertEquals("java.lang.Object", typeSpec.superclass.toString());
     }
 
     @Test
-    void givenPutEndpointBuilder_whenBuild_thenHaveMethodSpecIsCorrect() {
+    void givenPostEndpointBuilder_whenBuild_thenHaveMethodSpecIsCorrect() {
         MethodSpec methodSpec = typeSpec.methodSpecs.get(0);
         assertFalse(methodSpec.isConstructor());
-        assertEquals("put", methodSpec.name);
-
+        assertEquals("postParentChild", methodSpec.name);
         assertEquals(2, methodSpec.annotations.size());
+
         AnnotationSpec annotationSpec = methodSpec.annotations.get(0);
-        assertEquals("@org.springframework.web.bind.annotation.PutMapping(\"/{id}\")", annotationSpec.toString());
+        assertEquals("@org.springframework.web.bind.annotation.PostMapping", annotationSpec.toString());
         annotationSpec = methodSpec.annotations.get(1);
-        assertEquals("@org.springframework.web.bind.annotation.ResponseStatus(code = org.springframework.http.HttpStatus.OK)", annotationSpec.toString());
+        assertEquals("@org.springframework.web.bind.annotation.ResponseStatus(code = org.springframework.http.HttpStatus.CREATED)", annotationSpec.toString());
+
         assertEquals(1, methodSpec.modifiers.size());
         assertEquals("[public]", methodSpec.modifiers.toString());
 
-        assertEquals(3, methodSpec.parameters.size());
         ParameterSpec parameterSpec = methodSpec.parameters.get(0);
-        assertEquals("@org.springframework.web.bind.annotation.PathVariable(\"id\") java.lang.Long id", parameterSpec.toString());
+        assertEquals("[@org.springframework.web.bind.annotation.PathVariable(\"parent_id\")]", parameterSpec.annotations.toString());
         assertEquals("java.lang.Long", parameterSpec.type.toString());
-        assertEquals("id", parameterSpec.name);
+        assertEquals("parentId", parameterSpec.name);
 
         parameterSpec = methodSpec.parameters.get(1);
         assertEquals("[@org.springframework.web.bind.annotation.RequestBody, @javax.validation.Valid]", parameterSpec.annotations.toString());
-        assertEquals("the.group.artifact.entityname.web.UpdateEntityNameByIdResource", parameterSpec.type.toString());
+        assertEquals("the.group.artifact.child.web.CreateParentChildResource", parameterSpec.type.toString());
         assertEquals("body", parameterSpec.name);
 
-        parameterSpec = methodSpec.parameters.get(2);
-        assertEquals("[@org.springframework.web.bind.annotation.RequestAttribute]", parameterSpec.annotations.toString());
-        assertEquals("java.util.Set<java.lang.String>", parameterSpec.type.toString());
-        assertEquals("updatedFields", parameterSpec.name);
-
-        assertEquals("the.group.artifact.entityname.EntityName updateRequest = mapper.toEntity(body);\n" +
-                "service.update(id, updateRequest, updatedFields);\n" +
-                "the.group.artifact.entityname.EntityName createResult = service.search(id, null, null, null);\n" +
-                "the.group.artifact.entityname.web.EntityNameOutResource result = mapper.toResource(createResult);\n" +
-                "return new the.group.artifact.entityname.web.EntityNameResponse(result);\n", methodSpec.code.toString());
+        assertEquals("the.group.artifact.child.Child createRequest = mapper.toEntity(body);\n" +
+                "createRequest.setParent(new the.group.artifact.parent.Parent(parentId));\n" +
+                "service.create(createRequest);\n" +
+                "the.group.artifact.child.Child createResult = service.search(createRequest.getId(), null, null, null);\n" +
+                "the.group.artifact.child.web.ChildOutResource result = mapper.toResource(createResult);\n" +
+                "return new the.group.artifact.child.web.ChildResponse(result);\n", methodSpec.code.toString());
     }
 }