+++ generator-core/src/test/java/org/apiaddicts/apitools/apigen/generatorcore/generator/web/controller/endpoints/EndpointObjectMother.java
@@ -7,6 +7,7 @@ import org.apiaddicts.apitools.apigen.generatorcore.config.controller.Parameter;
 import org.apiaddicts.apitools.apigen.generatorcore.config.parameter.ParameterObjectMother;
 
 import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
 
 public class EndpointObjectMother {
@@ -23,6 +24,19 @@ public class EndpointObjectMother {
         return endpoint;
     }
 
+    public static Endpoint standardParentChildPost(String endpointName, String entityName) {
+        List<Parameter> parameters = new LinkedList<>();
+        Parameter parameter = ParameterObjectMother.getBasicPathParameter("parent_id", "integer");
+        parameters.add(parameter);
+        Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.POST, null, parameters);
+        endpoint.setRequest(EndpointRequestObjectMother.requestWithoutAttributes(entityName));
+        endpoint.setResponse(EndpointBaseResponseObjectMother.simpleResponseWithoutAttributes(entityName));
+        endpoint.setRelatedEntity(entityName);
+        endpoint.setParentEntity("Parent");
+        endpoint.setChildParentRelationProperty("parent.id");
+        return endpoint;
+    }
+
     public static Endpoint standardSearch(String endpointName, String entityName) {
         List<Parameter> parameters = ParameterObjectMother.createGetAllStandardParameters();
         Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.POST, "/search", parameters);
@@ -31,11 +45,35 @@ public class EndpointObjectMother {
         return endpoint;
     }
 
+    public static Endpoint standardParentChildSearch(String endpointName, String entityName) {
+        List<Parameter> parameters = ParameterObjectMother.createGetAllStandardParameters();
+        Parameter parameter = ParameterObjectMother.getBasicPathParameter("parent_id", "integer");
+        parameters.add(0, parameter);
+        Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.POST, "/search", parameters);
+        endpoint.setResponse(EndpointBaseResponseObjectMother.listResponseWithoutAttributesAndCollectionNameEqualsEntityName(entityName));
+        endpoint.setRelatedEntity(entityName);
+        endpoint.setParentEntity("Parent");
+        endpoint.setChildParentRelationProperty("parent.id");
+        return endpoint;
+    }
+
     public static Endpoint standardGetAll(String endpointName, String entityName) {
         List<Parameter> parameters = ParameterObjectMother.createGetAllStandardParameters();
         Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.GET, null, parameters);
         endpoint.setResponse(EndpointBaseResponseObjectMother.listResponseWithoutAttributesAndCollectionNameEqualsEntityName(entityName));
         endpoint.setRelatedEntity(entityName);
+        return endpoint;
+    }
+
+    public static Endpoint standardParentChildGetAll(String endpointName, String entityName) {
+        List<Parameter> parameters = ParameterObjectMother.createGetAllStandardParameters();
+        Parameter parameter = ParameterObjectMother.getBasicPathParameter("parent_id", "integer");
+        parameters.add(0, parameter);
+        Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.GET, null, parameters);
+        endpoint.setResponse(EndpointBaseResponseObjectMother.listResponseWithoutAttributesAndCollectionNameEqualsEntityName(entityName));
+        endpoint.setRelatedEntity(entityName);
+        endpoint.setParentEntity("Parent");
+        endpoint.setChildParentRelationProperty("parent.id");
         return endpoint;
     }
 
@@ -49,6 +87,20 @@ public class EndpointObjectMother {
         return endpoint;
     }
 
+    public static Endpoint standardParentChildGetById(String endpointName, String entityName) {
+        List<Parameter> parameters = ParameterObjectMother.createGetByIdStandardParameters();
+        Parameter parameter = ParameterObjectMother.getBasicPathParameter("parent_id", "integer");
+        parameters.add(0, parameter);
+        parameter = ParameterObjectMother.getBasicPathParameter("child_id", "integer");
+        parameters.add(1, parameter);
+        Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.GET, "/{child_id}", parameters);
+        endpoint.setResponse(EndpointBaseResponseObjectMother.simpleResponseWithoutAttributes(entityName));
+        endpoint.setRelatedEntity(entityName);
+        endpoint.setParentEntity("Parent");
+        endpoint.setChildParentRelationProperty("parent.id");
+        return endpoint;
+    }
+
     public static Endpoint standardPut(String endpointName, String entityName) {
         List<Parameter> parameters = new ArrayList<>();
         Parameter parameter = ParameterObjectMother.getBasicPathParameter("id", "integer");
@@ -57,7 +109,21 @@ public class EndpointObjectMother {
         endpoint.setRequest(EndpointRequestObjectMother.requestWithoutAttributes(entityName));
         endpoint.setResponse(EndpointBaseResponseObjectMother.simpleResponseWithoutAttributes(entityName));
         endpoint.setRelatedEntity(entityName);
-        ;
+        return endpoint;
+    }
+
+    public static Endpoint standardParentChildPut(String endpointName, String entityName) {
+        List<Parameter> parameters = new ArrayList<>();
+        Parameter parameter = ParameterObjectMother.getBasicPathParameter("parent_id", "integer");
+        parameters.add(parameter);
+        parameter = ParameterObjectMother.getBasicPathParameter("child_id", "integer");
+        parameters.add(parameter);
+        Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.PUT, "/{child_id}", parameters);
+        endpoint.setRequest(EndpointRequestObjectMother.requestWithoutAttributes(entityName));
+        endpoint.setResponse(EndpointBaseResponseObjectMother.simpleResponseWithoutAttributes(entityName));
+        endpoint.setRelatedEntity(entityName);
+        endpoint.setParentEntity("Parent");
+        endpoint.setChildParentRelationProperty("parent.id");
         return endpoint;
     }
 
@@ -68,6 +134,20 @@ public class EndpointObjectMother {
         Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.DELETE, "/{id}", parameters);
         endpoint.setResponse(EndpointBaseResponseObjectMother.simpleResponseWithoutAttributes(entityName));
         endpoint.setRelatedEntity(entityName);
+        return endpoint;
+    }
+
+    public static Endpoint standardParentChildDelete(String endpointName, String entityName) {
+        List<Parameter> parameters = new ArrayList<>();
+        Parameter parameter = ParameterObjectMother.getBasicPathParameter("parent_id", "integer");
+        parameters.add(parameter);
+        parameter = ParameterObjectMother.getBasicPathParameter("child_id", "integer");
+        parameters.add(parameter);
+        Endpoint endpoint = createEndpoint(endpointName, Endpoint.Method.DELETE, "/{child_id}", parameters);
+        endpoint.setResponse(EndpointBaseResponseObjectMother.simpleResponseWithoutAttributes(entityName));
+        endpoint.setRelatedEntity(entityName);
+        endpoint.setParentEntity("Parent");
+        endpoint.setChildParentRelationProperty("parent.id");
         return endpoint;
     }